requirements.txt
amqp==5.0.1
billiard==3.6.3.0
celery==5.0.0
Django==3.1.2
djangorestframework==3.12.1
kombu==5.0.2
Pillow==7.2.0
pytz==2020.1
redis==3.5.3
sqlparse==0.4.1
vine==5.0.0
psycopg2-binary==2.8.5
